launchpad: Refactor FeeTable type
