+++ packages/launchpad/types/signingcosmosclient.d.ts
@@ -1,6 +1,6 @@
 import { Coin } from "./coins";
 import { Account, BroadcastTxResult, CosmosClient, GetSequenceResult } from "./cosmosclient";
-import { GasLimits, GasPrice } from "./gas";
+import { FeeTable, GasLimits, GasPrice } from "./gas";
 import { BroadcastMode } from "./lcdapi";
 import { Msg } from "./msgs";
 import { StdFee } from "./types";
@@ -8,12 +8,12 @@ import { OfflineSigner } from "./wallet";
 /**
  * These fees are used by the higher level methods of SigningCosmosClient
  */
-export interface FeeTable extends Record<string, StdFee> {
+export interface CosmosFeeTable extends FeeTable {
   readonly send: StdFee;
 }
 /** Use for testing only */
 export interface PrivateSigningCosmosClient {
-  readonly fees: FeeTable;
+  readonly fees: CosmosFeeTable;
 }
 export declare class SigningCosmosClient extends CosmosClient {
   readonly senderAddress: string;
@@ -37,7 +37,7 @@ export declare class SigningCosmosClient extends CosmosClient {
     senderAddress: string,
     signer: OfflineSigner,
     gasPrice?: GasPrice,
-    gasLimits?: Partial<GasLimits<FeeTable>>,
+    gasLimits?: Partial<GasLimits<CosmosFeeTable>>,
     broadcastMode?: BroadcastMode,
   );
   getSequence(address?: string): Promise<GetSequenceResult>;
